+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=C:\Entwicklung\flutter"
+export "FLUTTER_APPLICATION_PATH=C:\Entwicklung\packages\get_it\example"
+export "FLUTTER_TARGET=lib\main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build\ios"
+export "FLUTTER_FRAMEWORK_DIR=C:\Entwicklung\flutter\bin\cache\artifacts\engine\ios"