+++ lib/get_it.dart
@@ -6,6 +6,8 @@ import 'package:meta/meta.dart';
 
 typedef FactoryFunc<T> = T Function();
 
+enum _readyStates { noneWaitingToBeReady, allReady, notAllReady }
+
 /// Very simple and easy to use service locator
 /// You register your object creation or an instance of an object with [registerFactory],
 /// [registerSingleton] or [registerLazySingleton]
@@ -54,7 +56,9 @@ class GetIt {
   /// retrieves or creates an instance of a registered type [T] depending on the registration function used for this type or based on a name.
   T get<T>([String instanceName]) {
     assert(!(!(const Object() is! T) && instanceName == null),
-        'You have to provide either a type or a name. Did you accidentally do  `var sl=GetIt.instance();` instead of var sl=GetIt.instance;');
+        'GetIt: You have to provide either a type or a name. Did you accidentally do  `var sl=GetIt.instance();` instead of var sl=GetIt.instance;');
+    assert(!(((const Object() is! T) && instanceName != null)),
+        'GetIt: You have to provide either a type OR a name not both.');
 
     _ServiceFactory<T> object;
     if (instanceName == null) {
@@ -81,40 +85,52 @@ class GetIt {
   /// registers a type so that a new instance will be created on each call of [get] on that type
   /// [T] type to register
   /// [func] factory function for this type
+  /// If [signalsReady] is set to `true` it means that the `ready` property cannot emit a ready event until this
+  /// registration was signalled ready
   /// [instanceName] if you provide a value here your factory gets registered with that
   /// name instead of a type. This should only be necessary if you need to register more
   /// than one instance of one type. Its highly not recommended
-  void registerFactory<T>(FactoryFunc<T> func, [String instanceName]) {
+  void registerFactory<T>(FactoryFunc<T> func,
+      {String instanceName, bool signalsReady = false}) {
     _register<T>(
         type: _ServiceFactoryType.alwaysNew,
         instanceName: instanceName,
-        factoryFunc: func);
+        factoryFunc: func,
+        signalsReady: signalsReady);
   }
 
   /// registers a type as Singleton by passing a factory function that will be called on the first call of [get] on that type
   /// [T] type to register
   /// [func] factory function for this type
+  /// If [signalsReady] is set to `true` it means that the `ready` property cannot emit a ready event until this
+  /// registration was signalled ready
   /// [instanceName] if you provide a value here your factory gets registered with that
   /// name instead of a type. This should only be necessary if you need to register more
   /// than one instance of one type. Its highly not recommended
-  void registerLazySingleton<T>(FactoryFunc<T> func, [String instanceName]) {
+  void registerLazySingleton<T>(FactoryFunc<T> func,
+      {String instanceName, bool signalsReady = false}) {
     _register<T>(
         type: _ServiceFactoryType.lazy,
         instanceName: instanceName,
-        factoryFunc: func);
+        factoryFunc: func,
+        signalsReady: signalsReady);
   }
 
   /// registers a type as Singleton by passing an [instance] of that type
   ///  that will be returned on each call of [get] on that type
   /// [T] type to register
+  /// If [signalsReady] is set to `true` it means that the `ready` property cannot emit a ready event until this
+  /// registration was signalled ready
   /// [instanceName] if you provide a value here your instance gets registered with that
   /// name instead of a type. This should only be necessary if you need to register more
   /// than one instance of one type. Its highly not recommended
-  void registerSingleton<T>(T instance, [String instanceName]) {
+  void registerSingleton<T>(T instance,
+      {String instanceName, bool signalsReady = false}) {
     _register<T>(
         type: _ServiceFactoryType.constant,
         instanceName: instanceName,
-        instance: instance);
+        instance: instance,
+        signalsReady: signalsReady);
   }
 
   /// Clears all registered types. Handy when writing unit tests
@@ -127,8 +143,10 @@ class GetIt {
       {@required _ServiceFactoryType type,
       FactoryFunc factoryFunc,
       T instance,
-      @required String instanceName}) {
-    assert(allowReassignment || !_factories.containsKey(T),
+      @required String instanceName,
+      @required bool signalsReady}) {
+    assert(
+        instanceName != null || allowReassignment || !_factories.containsKey(T),
         "Type ${T.toString()} is already registered");
     assert(
       instanceName != null ||
@@ -137,7 +155,9 @@ class GetIt {
     );
 
     var serviceFactory = _ServiceFactory<T>(type,
-        creationFunction: factoryFunc, instance: instance);
+        creationFunction: factoryFunc,
+        instance: instance,
+        shouldSignalReady: signalsReady);
     if (instanceName == null) {
       _factories[T] = serviceFactory;
     } else {
@@ -148,7 +168,10 @@ class GetIt {
   /// Unregister by Type [T] or by name [instanceName]
   /// if you need to dispose any resources you can do it using [disposingFunction] function
   /// that provides a instance of your class to be disposed
-  void unregister<T>({String instanceName, Function(T) disposingFunction}) {
+  void unregister<T>(
+      {String instanceName, void Function(T) disposingFunction}) {
+    assert(!(((const Object() is! T) && instanceName != null)),
+        'GetIt: You have to provide either a type OR a name not both.');
     assert(
         (instanceName != null && _factoriesByName.containsKey(instanceName)) ||
             _factories.containsKey(T),
@@ -162,21 +185,100 @@ class GetIt {
     }
   }
 
-  void signalReady() => _readySignalStream.add(true);
+  void signalReady<T>([String instanceName]) {
+    assert(!(((const Object() is! T) && instanceName != null)),
+        'GetIt.signalReady: You have to provide either a type OR a name not both.');
+    if ((const Object() is! T) ||
+        (instanceName != null && instanceName.isNotEmpty)) {
+      /// if (T is not a top level type especially not `dynamic`) or instanceName has a value
+      /// which means a specific registered object should be signalled
+      _ServiceFactory<T> object;
+      if (instanceName != null) {
+        object = _factoriesByName[instanceName];
+      } else if (const Object() is! T) {
+        object = _factories[T];
+      }
+      if (object == null) {
+        if (instanceName == null) {
+          throw Exception(
+              "GetIt.signalReady: Object of type ${T.toString()} is not registered inside GetIt");
+        } else {
+          throw Exception(
+              "GetIt.signalReady: Object with name $instanceName is not registered inside GetIt");
+        }
+      }
+      if (!object.shouldSignalReady) {
+        if (instanceName == null) {
+          throw Exception(
+              "GetIt.signalReady: Object of type ${T.toString()} does not wait to be signalled");
+        } else {
+          throw Exception(
+              "GetIt.signalReady: Object with name $instanceName does not wait to be signalled");
+        }
+      }
+      object.isReady = true;
+
+      if (_getReadyState() == _readyStates.allReady) {
+        _readySignalStream.add(true);
+      }
+    } else {
+      /// signalReady was called without a type or a name means a manual signalReady
+      var allReady = _getReadyState();
+      if (allReady == _readyStates.allReady ||
+          allReady == _readyStates.noneWaitingToBeReady) {
+        _readySignalStream.add(true);
+        return;
+      }
+
+      /// In case that there are still factories that are marked to wait for a signal
+      /// but aren't signalled we throw an exception with details which objects are concerned
+      final notReadyTypes = _factories.entries
+          .where((x) => (x.value.shouldSignalReady && !x.value.isReady))
+          .map<String>((x) => x.key.toString())
+          .toList();
+      final notReadyNames = _factoriesByName.entries
+          .where((x) => (x.value.shouldSignalReady && !x.value.isReady))
+          .map<String>((x) => x.key)
+          .toList();
+
+      throw (Exception(
+          'Registered types/names: $notReadyTypes  / $notReadyNames should signal ready but are not ready'));
+    }
+  }
+
+  _readyStates _getReadyState() {
+    var shouldbeSignalled = _factories.values
+        .where((x) => x.shouldSignalReady)
+        .toList()
+          ..addAll(_factoriesByName.values.where((x) => x.shouldSignalReady));
+    if (shouldbeSignalled.isEmpty) {
+      return _readyStates.noneWaitingToBeReady;
+    }
+    if (shouldbeSignalled.every((x) => x.isReady)) {
+      return _readyStates.allReady;
+    }
+    return _readyStates.notAllReady;
+  }
 }
 
 enum _ServiceFactoryType { alwaysNew, constant, lazy }
 
 class _ServiceFactory<T> {
-  _ServiceFactoryType type;
-  FactoryFunc creationFunction;
+  final _ServiceFactoryType factoryType;
+  final FactoryFunc creationFunction;
   Object instance;
+  final bool shouldSignalReady;
+  bool isReady;
 
-  _ServiceFactory(this.type, {this.creationFunction, this.instance});
+  _ServiceFactory(this.factoryType,
+      {this.creationFunction,
+      this.instance,
+      this.isReady = false,
+      this.shouldSignalReady = false});
 
   T getObject() {
     try {
-      switch (type) {
+      switch (factoryType) {
         case _ServiceFactoryType.alwaysNew:
           return creationFunction() as T;
           break;
@@ -195,6 +297,5 @@ class _ServiceFactory<T> {
       print('Stack trace:\n $s');
       rethrow;
     }
-    return null; // should never get here but to make the analyser happy
   }
 }